fork and clone, migrations, deprecations
The first instruction says to fork and clone the fork, but then the given command is for cloning directly. I replaced saasbook with your_github_username. Next, I added a line about ignoring deprecation warnings generated by using an old version of rspec-rails, discussed in two of the issues threads. One fix seems to be to remove the version from rspec-rails in the gemfile (https://github.com/rspec/rspec-core/issues/2210) but it should be safe to ignore for now. Finally, running cucumber before rake db:test:prepare causes a notice of pending migrations and a suggestion of rake db:migrate RAILS_ENV=test, which I believe accomplishes the same thing.